+++ src/eoyeon/Student.java
@@ -1,4 +1,4 @@
-package sample;
+package eoyeon;
 
 import java.time.LocalDate;
