+++ src/eoyeon/StudentDetail.java
@@ -1,4 +1,4 @@
-package sample;
+package eoyeon;
 
 import java.util.List;
 import java.util.Optional;
