--- src/sample/StudentDetailImpl.java
@@ -1,50 +1,0 @@
-package sample;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-public class StudentDetailImpl implements StudentDetail {
-    Map<Integer, Student> studentsMap = new HashMap<>();
-
-    @Override
-    public boolean addStudent(int id, Student student) {
-
-    }
-
-    @Override
-    public boolean removeStudent(int id) {
-
-    }
-
-    @Override
-    public boolean updateStudent(int id, Student student) {
-
-    }
-
-    @Override
-    public int getStudentCount() {
-
-    }
-
-    @Override
-    public List<String> findStudentsByAgeMoreThan(int age) {
-
-    }
-
-    @Override
-    public Optional<Student> findStudentByPart(String part) {
-
-    }
-
-    @Override
-    public List<Student> findAllStudents() {
-
-    }
-
-    @Override
-    public void printStudents() {
-
-    }
-}
